requirements.txt
plotly==5.17.0
pandas==2.1.2
numpy==1.26.0
scikit-learn == 1.3.1
lightweight-charts==1.0.17.6
streamlit-lightweight-charts==0.7.20
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.5
pyarrow==13.0.0
torch==2.0.1
pyyaml==6.0.1
torchvision==0.16.0
matplotlib==3.8.0
mplfinance==0.12.10b0